Added check to ensure we don't have a shortage of tokens to supply presale contributors

+++ contracts/DNNToken.sol
@@ -308,7 +308,7 @@ contract DNNToken is StandardToken {
     //////////////////
     // Issue tokens //
     //////////////////
-    function issueTokens(address beneficiary, uint tokenCount, DNNSupplyAllocations allocationType)
+    function issueTokens(address beneficiary, uint256 tokenCount, DNNSupplyAllocations allocationType)
       onlyAllocatorOrCrowdfundContractOrPlatform
       returns (bool)
     {
